fixed issue with mysql injection and link protection

+++ admin/includes/components/view_all_comments.php
@@ -47,8 +47,10 @@ while ($row = mysqli_fetch_assoc($resultAllComments)) {
 }
 
 if(isset($_GET['delete'])) {
-
+    //validation if the user has authorization to delete user's
+  if (isset($_SESSION['user_role']) && $_SESSION['user_role' == 'administrator']) {  
     deleteComment();
+    }
 }
 
 if(isset($_GET['update'])) {
